[watsonx.ai] Refactor WatsonxChatModel/WatsonxStreamingChatModel to WatsonxGenerationModel
