+++ model-providers/watsonx/deployment/src/test/java/io/quarkiverse/langchain4j/watsonx/deployment/PromptFormatterTest.java
@@ -127,6 +127,7 @@ public class PromptFormatterTest extends WireMockAbstract {
                 .temperature(chatModelConfig.temperature())
                 .minNewTokens(chatModelConfig.minNewTokens())
                 .maxNewTokens(chatModelConfig.maxNewTokens())
+                .timeLimit(10000L)
                 .build();
 
         TextGenerationRequest body = new TextGenerationRequest(modelId, projectId,
