+++ model-providers/watsonx/runtime/src/main/java/io/quarkiverse/langchain4j/watsonx/runtime/WatsonxRecorder.java
@@ -20,11 +20,9 @@ import dev.langchain4j.model.chat.StreamingChatLanguageModel;
 import dev.langchain4j.model.embedding.DisabledEmbeddingModel;
 import dev.langchain4j.model.embedding.EmbeddingModel;
 import io.quarkiverse.langchain4j.runtime.NamedConfigUtil;
-import io.quarkiverse.langchain4j.watsonx.TokenGenerator;
-import io.quarkiverse.langchain4j.watsonx.WatsonxChatModel;
 import io.quarkiverse.langchain4j.watsonx.WatsonxEmbeddingModel;
-import io.quarkiverse.langchain4j.watsonx.WatsonxModel;
-import io.quarkiverse.langchain4j.watsonx.WatsonxStreamingChatModel;
+import io.quarkiverse.langchain4j.watsonx.WatsonxGenerationModel;
+import io.quarkiverse.langchain4j.watsonx.WatsonxTokenGenerator;
 import io.quarkiverse.langchain4j.watsonx.prompt.PromptFormatter;
 import io.quarkiverse.langchain4j.watsonx.prompt.impl.NoopPromptFormatter;
 import io.quarkiverse.langchain4j.watsonx.runtime.config.ChatModelConfig;
@@ -43,10 +41,10 @@ public class WatsonxRecorder {
     private static final String DUMMY_URL = "https://dummy.ai/api";
     private static final String DUMMY_API_KEY = "dummy";
     private static final String DUMMY_PROJECT_ID = "dummy";
-    private static final Map<String, TokenGenerator> tokenGeneratorCache = new HashMap<>();
+    private static final Map<String, WatsonxTokenGenerator> tokenGeneratorCache = new HashMap<>();
     private static final ConfigValidationException.Problem[] EMPTY_PROBLEMS = new ConfigValidationException.Problem[0];
 
-    public Supplier<ChatLanguageModel> chatModel(LangChain4jWatsonxConfig runtimeConfig,
+    public Supplier<ChatLanguageModel> generationModel(LangChain4jWatsonxConfig runtimeConfig,
             LangChain4jWatsonxFixedRuntimeConfig fixedRuntimeConfig,
             String configName, PromptFormatter promptFormatter) {
 
@@ -65,7 +63,7 @@ public class WatsonxRecorder {
             return new Supplier<>() {
                 @Override
                 public ChatLanguageModel get() {
-                    return builder.build(WatsonxChatModel.class);
+                    return builder.build();
                 }
             };
 
@@ -79,7 +77,7 @@ public class WatsonxRecorder {
         }
     }
 
-    public Supplier<StreamingChatLanguageModel> streamingChatModel(LangChain4jWatsonxConfig runtimeConfig,
+    public Supplier<StreamingChatLanguageModel> generationStreamingModel(LangChain4jWatsonxConfig runtimeConfig,
             LangChain4jWatsonxFixedRuntimeConfig fixedRuntimeConfig, String configName, PromptFormatter promptFormatter) {
 
         LangChain4jWatsonxConfig.WatsonConfig watsonRuntimeConfig = correspondingWatsonRuntimeConfig(runtimeConfig, configName);
@@ -92,7 +90,7 @@ public class WatsonxRecorder {
             return new Supplier<>() {
                 @Override
                 public StreamingChatLanguageModel get() {
-                    return builder.build(WatsonxStreamingChatModel.class);
+                    return builder.build();
                 }
             };
 
@@ -117,7 +115,7 @@ public class WatsonxRecorder {
             }
 
             String iamUrl = watsonConfig.iam().baseUrl().toExternalForm();
-            TokenGenerator tokenGenerator = tokenGeneratorCache.computeIfAbsent(iamUrl,
+            WatsonxTokenGenerator tokenGenerator = tokenGeneratorCache.computeIfAbsent(iamUrl,
                     createTokenGenerator(watsonConfig.iam(), watsonConfig.apiKey()));
 
             URL url;
@@ -141,7 +139,7 @@ public class WatsonxRecorder {
             return new Supplier<>() {
                 @Override
                 public WatsonxEmbeddingModel get() {
-                    return builder.build(WatsonxEmbeddingModel.class);
+                    return builder.build();
                 }
             };
 
@@ -155,18 +153,18 @@ public class WatsonxRecorder {
         }
     }
 
-    private Function<? super String, ? extends TokenGenerator> createTokenGenerator(IAMConfig iamConfig, String apiKey) {
-        return new Function<String, TokenGenerator>() {
+    private Function<? super String, ? extends WatsonxTokenGenerator> createTokenGenerator(IAMConfig iamConfig, String apiKey) {
+        return new Function<String, WatsonxTokenGenerator>() {
 
             @Override
-            public TokenGenerator apply(String iamUrl) {
-                return new TokenGenerator(iamConfig.baseUrl(), iamConfig.timeout().orElse(Duration.ofSeconds(10)),
+            public WatsonxTokenGenerator apply(String iamUrl) {
+                return new WatsonxTokenGenerator(iamConfig.baseUrl(), iamConfig.timeout().orElse(Duration.ofSeconds(10)),
                         iamConfig.grantType(), apiKey);
             }
         };
     }
 
-    private WatsonxModel.Builder generateChatBuilder(
+    private WatsonxGenerationModel.Builder generateChatBuilder(
             LangChain4jWatsonxConfig.WatsonConfig watsonRuntimeConfig,
             LangChain4jWatsonxFixedRuntimeConfig.WatsonConfig watsonFixedRuntimeConfig,
             String configName, PromptFormatter promptFormatter) {
@@ -179,7 +177,7 @@ public class WatsonxRecorder {
         }
 
         String iamUrl = watsonRuntimeConfig.iam().baseUrl().toExternalForm();
-        TokenGenerator tokenGenerator = tokenGeneratorCache.computeIfAbsent(iamUrl,
+        WatsonxTokenGenerator tokenGenerator = tokenGeneratorCache.computeIfAbsent(iamUrl,
                 createTokenGenerator(watsonRuntimeConfig.iam(), watsonRuntimeConfig.apiKey()));
 
         URL url;
@@ -193,7 +191,7 @@ public class WatsonxRecorder {
         Integer startIndex = chatModelConfig.lengthPenalty().startIndex().orElse(null);
         String promptJoiner = chatModelConfig.promptJoiner();
 
-        return WatsonxChatModel.builder()
+        return WatsonxGenerationModel.builder()
                 .tokenGenerator(tokenGenerator)
                 .url(url)
                 .timeout(watsonRuntimeConfig.timeout().orElse(Duration.ofSeconds(10)))
